third_party/WebKit/LayoutTests/media/video-enter-fullscreen-without-user-gesture.html

<html>
    <head>
        <title>Test that video webkitRequestFullscreen() works without any user gesture if the requirement is removed.</title>
        <script src=media-controls.js></script>
        <script src=media-file.js></script>
        <script src=video-test.js></script>
        <script>
            if (window.internals)
                window.internals.settings.setMediaFullscreenRequiresUserGesture(false);

            function canplaythrough()
            {
                consoleWrite("");
                consoleWrite("* No user gesture initiated");
                testExpected("video.paused", true);
                run("video.webkitRequestFullscreen()");
                waitForEvent('webkitfullscreenchange', fullscreenchange);
            }

            function fullscreenchange()
            {
                endTest();
            }

            function start()
            {
                 findMediaElement();
                 waitForEvent('canplaythrough', canplaythrough);
                 video.src = findMediaFile("video", "content/test");
            }
        </script>
    </head>

    <body onload="start()">
    <p>Test that video webkitRequestFullscreen() works without any user gesture if the requirement is removed.</p>
    <video controls></video>
    </body>
</html>
